refactor: Relocate remote DTOs to model package
