+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/remote/mapper/prayertime/prayerTimesDtoMapper.kt
@@ -1,7 +1,7 @@
 package net.thechance.mena.faith.data.remote.mapper.prayertime
 
 import kotlinx.datetime.TimeZone
-import net.thechance.mena.faith.data.remote.dto.prayertime.PrayerTimes
+import net.thechance.mena.faith.data.remote.model.prayertime.PrayerTimes
 import net.thechance.mena.faith.domain.entity.PrayerName
 import net.thechance.mena.faith.domain.entity.PrayerTime
 import kotlin.time.ExperimentalTime
