+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/remote/model/prayertime/PrayerDateDto.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.faith.data.remote.dto.prayertime
+package net.thechance.mena.faith.data.remote.model.prayertime
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
