+++ faith_data/src/commonTest/kotlin/net/thechance/mena/faith/data/repository/prayerTimesFakeData.kt
@@ -3,10 +3,10 @@ package net.thechance.mena.faith.data.repository
 import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toInstant
-import net.thechance.mena.faith.data.remote.dto.prayertime.GregorianDate
-import net.thechance.mena.faith.data.remote.dto.prayertime.HijriDate
-import net.thechance.mena.faith.data.remote.dto.prayertime.PrayerDate
-import net.thechance.mena.faith.data.remote.dto.prayertime.PrayerTimes
+import net.thechance.mena.faith.data.remote.model.prayertime.GregorianDate
+import net.thechance.mena.faith.data.remote.model.prayertime.HijriDate
+import net.thechance.mena.faith.data.remote.model.prayertime.PrayerDate
+import net.thechance.mena.faith.data.remote.model.prayertime.PrayerTimes
 import net.thechance.mena.faith.domain.entity.PrayerName
 import net.thechance.mena.faith.domain.entity.PrayerTime
 import kotlin.time.ExperimentalTime
